Enable error handling from main to renderer
Apparently you cant pass thrown erros from main to renderer so this is
the best I can come up with. The preload script catches the error and
all ipc responses have an optional data and error field. However this
relies on the renderer manually checking if an error was returned